Support generating certificates based on CSR files and add some practical alias (#3)
* Support generating certificates based on CSR files

* add alias: certm-cdroot, certm-cdclients, certm-cdservers, certm-cdca, certm-lsclients, certm-lsservers, certm-lsca

+++ src/tools/mkcert.sh
@@ -11,11 +11,15 @@ conf_domain_name=
 conf_gm_enable=
 conf_type=clients
 conf_cert_type=rsa
+
+conf_csr=
+conf_key=
 conf_begin=
 conf_end=
 
 cert_dir=
 date_options=
+g_clean_cert_dir=yes
 
 ####################### Functions #######################
 
@@ -28,6 +32,8 @@ function usage
     echo "  -e, --end   <DATE>                      End date, default is 1095 days"
     echo "  -g, --gm                                GM certificate (deprecated, use \"-t sm2\" instead)"
     echo "  -h, --help                              Show help"
+    echo "  -k, --key <PRIVATE_KEY_FILE>            Private key file. If specified of CSR file(-r), will use this key file"
+    echo "  -r, --request <CSR_FILE>                CSR file. If specified, will make certificate from CSR file"
     echo "  -s, --server                            Server certificate, default is client"
     echo "  -t, --type  <rsa | ecdsa | sm2>         Certificate Key type, default is 'rsa'"
     echo ""
@@ -67,12 +73,14 @@ function exit_on_error
 
     cd -
 
-    rm -rf $cert_dir
+    if [ "$g_clean_cert_dir" == "yes" ]; then
+        rm -rf $cert_dir
 
-    domain_dir=`dirname $cert_dir`
+        domain_dir=`dirname $cert_dir`
 
-    if [ -z "`ls -A $domain_dir`" ]; then
-        rm -rf $domain_dir
+        if [ -z "`ls -A $domain_dir`" ]; then
+            rm -rf $domain_dir
+        fi
     fi
 
     exit 1;
@@ -94,20 +102,27 @@ function convert_p12
 
 function gen_rsa
 {
-    # gen key
+    gencsr=$1
 
-    $CERTM_OPENSSL genpkey -out $cert_dir/privkey.pem \
-                           -algorithm RSA \
-                           -pkeyopt rsa_keygen_bits:2048
-    [ $? -eq 0 ] || exit_on_error
+    if [ "$gencsr" == "yes" ]; then
+        # gen key
 
-    # gen csr
+        $CERTM_OPENSSL genpkey -out $conf_key \
+                            -algorithm RSA \
+                            -pkeyopt rsa_keygen_bits:2048
+        [ $? -eq 0 ] || exit_on_error
 
-    $CERTM_OPENSSL req -new \
-                       -config $cert_dir/csr.conf \
-                       -key $cert_dir/privkey.pem \
-                       -out $cert_dir/priv.csr
-    [ $? -eq 0 ] || exit_on_error
+        # gen csr
+        $CERTM_OPENSSL req -new \
+                        -config $cert_dir/csr.conf \
+                        -key $conf_key \
+                        -out $conf_csr
+        [ $? -eq 0 ] || exit_on_error
+    else
+        # check key type wthether is RSA
+        pkey_get_type $conf_key | grep -q rsa \
+            || { echo "Key type must be RSA"; exit 1; }
+    fi
 
     # gen cert
 
@@ -115,7 +130,7 @@ function gen_rsa
 
     $CERTM_OPENSSL ca -config ca.conf \
                       $date_options \
-                      -in $cert_dir/priv.csr \
+                      -in $conf_csr \
                       -out $cert_dir/cert.pem \
                       -extensions server_ext \
                       -notext \
@@ -137,26 +152,34 @@ function gen_rsa
     [ $? -eq 0 ] || exit_on_error
 
     # gen p12
-    convert_p12 $cert_dir/cert.pem $cert_dir/privkey.pem $cert_dir/cert.p12
+    convert_p12 $cert_dir/cert.pem $conf_key $cert_dir/cert.p12
     [ $? -eq 0 ] || exit_on_error
 }
 
 function gen_ecdsa
 {
-    # gen key
+    gencsr=$1
 
-    $CERTM_OPENSSL genpkey -out $cert_dir/privkey.pem \
-                           -algorithm EC \
-                           -pkeyopt ec_paramgen_curve:P-256
-    [ $? -eq 0 ] || exit_on_error
+    if [ "$gencsr" == "yes" ]; then
+        # gen key
 
-    # gen csr
+        $CERTM_OPENSSL genpkey -out $conf_key \
+                            -algorithm EC \
+                            -pkeyopt ec_paramgen_curve:P-256
+        [ $? -eq 0 ] || exit_on_error
 
-    $CERTM_OPENSSL req -new \
-                       -config $cert_dir/csr.conf \
-                       -key $cert_dir/privkey.pem \
-                       -out $cert_dir/priv.csr
-    [ $? -eq 0 ] || exit_on_error
+        # gen csr
+
+        $CERTM_OPENSSL req -new \
+                        -config $cert_dir/csr.conf \
+                        -key $conf_key \
+                        -out $conf_csr
+        [ $? -eq 0 ] || exit_on_error
+    else
+        # check key type wthether is ecdsa
+        pkey_get_type $conf_key | grep -q ecdsa \
+            || { echo "Key type must be RSA"; exit 1; }
+    fi
 
     # gen cert
 
@@ -164,7 +187,7 @@ function gen_ecdsa
 
     $CERTM_OPENSSL ca -config ca.conf \
                       $date_options \
-                      -in $cert_dir/priv.csr \
+                      -in $conf_csr \
                       -out $cert_dir/cert.pem \
                       -extensions server_ext \
                       -notext \
@@ -186,7 +209,7 @@ function gen_ecdsa
     [ $? -eq 0 ] || exit_on_error
 
     # gen p12
-    convert_p12 $cert_dir/cert.pem $cert_dir/privkey.pem $cert_dir/cert.p12
+    convert_p12 $cert_dir/cert.pem $conf_key $cert_dir/cert.p12
     [ $? -eq 0 ] || exit_on_error
 }
 
@@ -283,6 +306,73 @@ function gen_gm
     [ $? -eq 0 ] || exit_on_error
 }
 
+# check csr and key file is matched
+function check_csr
+{
+    csr=$1
+    key=$2
+    openssl_conf=$CERTM_SUB_CA_DIR/ca.conf
+
+    [ -z "$conf_csr" ] && return 0
+
+    # check csr file is existed
+    if [ ! -f "$conf_csr" ];then
+        echo "CSR file $conf_csr is not existed"
+        return 1
+    fi
+
+    # check key file is existed
+    [ -z "$conf_key" ] \
+        && echo "private key must be specified, if csr file(-r) is specified" \
+        && return 1
+
+    if [ ! -f "$conf_key" ];then
+        echo "Key file $conf_key is not existed"
+        return 1
+    fi
+
+    t=`mktemp`
+    if $CERTM_OPENSSL req \
+        -config $openssl_conf \
+        -in $conf_csr \
+        -key $conf_key \
+        -noout \
+        -verify \
+        2>&1 | tee $t | grep -q "verify OK";
+    then
+        rm $t
+        return 0
+    fi
+
+    echo "csr error: `cat $t`"
+    rm $t;
+
+    return 1
+}
+
+function pkey_get_type
+{
+    key=$1
+
+    # checking private key whether is existed
+    if [ ! -f "$key" ]; then
+        echo "Private key file ($key) not found."
+        exit 1
+    fi
+
+    # get information of private key
+    info=$(openssl pkey -in $key -text -noout 2>&1)
+
+    # judge the type of private key
+    if echo "$info" | grep -q "Private-Key: (.* bit, .* primes)"; then
+        echo "rsa"
+    elif echo "$info" | grep -q "prime256v1"; then
+        echo "ecdsa"
+    else
+        echo "unknown"
+    fi
+}
+
 ####################### Main #######################
 
 trap exit_on_error SIGINT
@@ -292,6 +382,8 @@ trap exit_on_error SIGINT
 if [ $# -eq 0 ]; then
     usage
 fi
+
+opt_gencsr=yes
 
 while [ $# -gt 0 ]
 do
@@ -312,6 +404,16 @@ do
             ;;
         -h|--help)
             usage
+            ;;
+        -k|--key)
+            conf_key=$2
+            shift
+            ;;
+        -r|--request)
+            conf_csr=$2
+            opt_gencsr=no
+            g_clean_cert_dir=no
+            shift;
             ;;
         -s|--server)
             conf_type=servers
@@ -338,28 +440,46 @@ do
     shift
 done
 
+if [ -n "$conf_csr" ]; then
+    check_csr $conf_csr $conf_key \
+        || { echo "CSR '$conf_csr' and Key '$conf_key' is not matched"; exit 1; }
 
-if [ -z "$conf_domain_name" ]; then
-    echo "Must specify domain name"
-    usage
-fi
-
-dn=$conf_domain_name.$g_conf_domain_suffix
-
-if [ "$conf_type" == "servers" ]; then
-    cert_dir=$CERTM_SERVER_DIR/$dn/$conf_cert_type
+    cert_dir=`dirname $conf_csr`
 else
-    cert_dir=$CERTM_CLIENT_DIR/$dn/$conf_cert_type
+    # check config
+    if [ -z "$conf_domain_name" ]; then
+        echo "Must specify domain name"
+        usage
+    fi
+
+    dn=$conf_domain_name.$g_conf_domain_suffix
+
+    if [ "$conf_type" == "servers" ]; then
+        cert_dir=$CERTM_SERVER_DIR/$dn/$conf_cert_type
+    else
+        cert_dir=$CERTM_CLIENT_DIR/$dn/$conf_cert_type
+    fi
+
+    # mkdir directory
+
+    if [ -d $cert_dir ]; then
+        echo "$cert_dir was existed"
+        exit 0
+    fi
+
+    mkdir -p $cert_dir
+
+    # init csr
+
+    cp $CERTM_CSR_CONF $cert_dir/csr.conf
+    sed -i "s/{{domain_name}}/$dn/g" $cert_dir/csr.conf
+    sed -i "s/{{organization}}/$g_conf_organization/g" $cert_dir/csr.conf
+    sed -i "s/{{organization_unit}}/$g_conf_organization_unit/g" $cert_dir/csr.conf
+
+    # init default config
+    conf_csr=$cert_dir/priv.csr
+    conf_key=$cert_dir/privkey.pem
 fi
-
-# mkdir directory
-
-if [ -d $cert_dir ]; then
-    echo "$cert_dir was existed"
-    exit 0
-fi
-
-mkdir -p $cert_dir
 
 # validate default is +1095 days, else is from $conf_begin to $conf_end
 if [ -n "$conf_begin" -a -n "$conf_end" ]; then
@@ -372,24 +492,21 @@ else
     date_options="-days 1095"
 fi
 
-# init config
-
-cp $CERTM_CSR_CONF $cert_dir/csr.conf
-sed -i "s/{{domain_name}}/$dn/g" $cert_dir/csr.conf
-sed -i "s/{{organization}}/$g_conf_organization/g" $cert_dir/csr.conf
-sed -i "s/{{organization_unit}}/$g_conf_organization_unit/g" $cert_dir/csr.conf
-
 case $conf_cert_type in
     rsa)
-        gen_rsa
+        gen_rsa $opt_gencsr
         ;;
     ecdsa)
-        gen_ecdsa
+        gen_ecdsa $opt_gencsr
         ;;
     sm2)
         conf_gm_enable=1
 
-        # enc
+        # TODO: don't support csr now
+        [ "$opt_gencsr" == "no" ] \
+            && { echo "SM2 certificate don't support CSR"; exit 1; }
+
+        # init csr enc
         cp $CERTM_ENC_CSR_CONF $cert_dir/enc-csr.conf
         sed -i "s/{{domain_name}}/$dn/g" $cert_dir/enc-csr.conf
         sed -i "s/{{organization}}/$g_conf_organization/g" $cert_dir/enc-csr.conf
